requirements.txt
pandas>=2.1
pyarrow>=14.0
PyYAML>=6.0
requests>=2.31
python-dateutil>=2.9